temp3

+++ domain/src/main/kotlin/com/g2s/trading/indicator/IndicatorUseCase.kt
@@ -5,6 +5,7 @@ import com.g2s.trading.order.Symbol
 import com.g2s.trading.indicator.indicator.CandleStick
 import com.g2s.trading.indicator.indicator.Interval
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 
 @Service
 class IndicatorUseCase(
@@ -15,7 +16,7 @@ class IndicatorUseCase(
         return exchangeImpl.getCandleStick(symbol, interval, limit)
     }
 
-    fun getLastPrice(symbol: Symbol): Double {
-        return exchangeImpl.getLastPrice(symbol)
+    fun getLastPrice(symbol: Symbol): BigDecimal {
+        return BigDecimal(exchangeImpl.getLastPrice(symbol))
     }
 }
